+++ packages/bpk-docs/src/pages/ChipsPage/index.js
@@ -1,0 +1,3 @@
+import page from './ChipsPage';
+
+export default page;